+++ R/data.R
@@ -11,6 +11,7 @@
 #'   \item{vaccine}{binary indicator of treatment arm, with value 1 in vaccinated and 0 in control subjects}
 #'   \item{type_disease}{serotype of disease}
 #'   \item{disease_any}{binary indicator of disease caused by any serotype}
+#'   \item{time_event}{time to disease or end of follow-up in days}
 #' }
 "data_temp"