+++ tests/testthat.R
@@ -2,6 +2,7 @@ library(testthat)
 library(vaxpmx)
 library(dplyr)
 library(MASS)
+library(survival)
 
 # test_dir("tests/testthat")
 test_check("vaxpmx")