+++ NEWS.md
@@ -1,10 +1,15 @@
 # Package versions
 
 ## 0.0.1 version
-Initial version of the package, proprietary
+Initial version of the package, proprietary.
 
 ## 0.0.2 version
-Version with logistic regression
+Version with logistic regression.
 
 ## 0.0.3 version
-Version to be uploaded to CRAN, accompanies the publication by Dudasova et al., 2024, BMC Med Res Methodol
+Version to be uploaded to CRAN, accompanies the publication by Dudasova, Valenta & Sachs, 2024, BMC Med Res Methodol.
+
+## 0.0.4 version
+## 0.0.5 version
+Version to be uploaded to CRAN, accompanies the publication by Dudasova, Valenta & Sachs, 2024, NPJ Vaccines.
+Implementation of efficacy estimation using time-to-event models (fitted with coxph() function).